@@ -14,7 +14,7 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.2.2 <3.0.0"
+  sdk: ">=2.17.0-69.0.dev <3.0.0"
 
 dependencies:
   flutter:
@@ -25,6 +25,8 @@ dependencies:
   cupertino_icons: ^0.1.3
   responsive_builder: ^0.2.0+2
   get_it: ^4.0.2
+  fl_chart: ^0.51.0
+  charts_flutter: ^0.12.0
 
 dev_dependencies:
   flutter_test:
