@@ -2,13 +2,13 @@ import 'package:flutter/material.dart';
 
 class CenteredView extends StatelessWidget {
   final Widget child;
-  const CenteredView({Key key, this.child}) : super(key: key);
+  const CenteredView({Key? key, required this.child}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     return Container(
       padding: const EdgeInsets.symmetric(horizontal: 70, vertical: 60),
-      alignment: Alignment.topCenter,
+      alignment: Alignment.topLeft,
       child: ConstrainedBox(
         constraints: BoxConstraints(maxWidth: 1200),
         child: child,
