@@ -1,16 +1,45 @@
 import 'package:flutter/material.dart';
-import 'package:the_basics/widgets/navigation_bar/navigation_bar_tablet_desktop.dart';
-import 'package:the_basics/widgets/navigation_bar/navigation_bar_mobile.dart';
-import 'package:responsive_builder/responsive_builder.dart';
 
 class NavigationBar extends StatelessWidget {
-  const NavigationBar({Key key}) : super(key: key);
+  const NavigationBar({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    return ScreenTypeLayout(
-      mobile: NavigationBarMobile(),
-      tablet: NavigationBarTabletDesktop(),
+    return Container(
+      height: 100,
+      child: Row(
+        mainAxisAlignment: MainAxisAlignment.spaceBetween,
+        children: <Widget>[
+          // SizedBox(
+          //   height: 80,
+          //   width: 150,
+          //   child: Image.asset('assets/logo.png'),
+          // ),
+          Row(
+            mainAxisSize: MainAxisSize.min,
+            children: <Widget>[
+              _NavBarItem('Episodes'),
+              SizedBox(
+                width: 60,
+              ),
+              _NavBarItem('About'),
+            ],
+          )
+        ],
+      ),
+    );
+  }
+}
+
+class _NavBarItem extends StatelessWidget {
+  final String title;
+  const _NavBarItem(this.title);
+
+  @override
+  Widget build(BuildContext context) {
+    return Text(
+      title,
+      style: TextStyle(fontSize: 18),
     );
   }
 }
