@@ -1,12 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:the_basics/views/layout_template/layout_template.dart';
+import 'package:the_basics/views/home/home_view.dart';
 
-import 'locator.dart';
-
-void main() {
-  setupLocator();
-  runApp(MyApp());
-}
+void main() => runApp(MyApp());
 
 class MyApp extends StatelessWidget {
   @override
@@ -17,7 +12,7 @@ class MyApp extends StatelessWidget {
         primarySwatch: Colors.blue,
         textTheme: Theme.of(context).textTheme.apply(fontFamily: 'Open Sans'),
       ),
-      home: LayoutTemplate(),
+      home: HomeView(),
     );
   }
 }
