@@ -1,7 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:responsive_builder/responsive_builder.dart';
-import 'package:the_basics/widgets/call_to_action/call_to_action_mobile.dart';
-import 'package:the_basics/widgets/call_to_action/call_to_action_tablet_desktop.dart';
 
 class CallToAction extends StatelessWidget {
   final String title;
@@ -9,9 +6,20 @@ class CallToAction extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return ScreenTypeLayout(
-      mobile: CallToActionMobile(title),
-      tablet: CallToActionTabletDesktop(title),
+    return Container(
+      padding: const EdgeInsets.symmetric(horizontal: 60, vertical: 15),
+      child: Text(
+        title,
+        style: TextStyle(
+          fontSize: 18,
+          fontWeight: FontWeight.w800,
+          color: Colors.white,
+        ),
+      ),
+      decoration: BoxDecoration(
+        color: Color.fromARGB(255, 31, 229, 146),
+        borderRadius: BorderRadius.circular(5),
+      ),
     );
   }
 }
